pom: don't activate jdk_unix profile on mac
This profile was being activated on Mac, even though that wasn't the intention.
This can be checked by running "mvn help:active-profiles".  The cause appears
to have been reported as MNG-4983.  The fix in this case is to explicitly
declare that for our purposes, we exclude mac from the unix profile.

http://jira.codehaus.org/browse/MNG-4983